+++ Changelog.md
@@ -1,5 +1,6 @@
 ## 0.0.41
 + Fixed invalid data type error when prediction returns a greyscale image.
++ Refactored `Acceleration.Default` enumeration member to `Acceleration.Auto`.
 
 ## 0.0.40
 + Fixed errors when `Function` client is created for the first time on a new device.